+++ x-pack/plugins/infra/public/components/alerting/inventory/expression.tsx
@@ -336,6 +336,10 @@ export const Expressions: React.FC<Props> = (props) => {
   );
 };
 
+// required for dynamic import
+// eslint-disable-next-line import/no-default-export
+export default Expressions;
+
 interface ExpressionRowProps {
   nodeType: InventoryItemType;
   expressionId: number;
