+++ x-pack/plugins/infra/public/components/alerting/logs/expression_editor/editor.tsx
@@ -236,3 +236,7 @@ export const Editor: React.FC<Props> = (props) => {
     </>
   );
 };
+
+// required for dynamic import
+// eslint-disable-next-line import/no-default-export
+export default Editor;
