+++ x-pack/plugins/infra/public/components/alerting/logs/log_threshold_alert_type.ts
@@ -4,10 +4,10 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import { i18n } from '@kbn/i18n';
+import React from 'react';
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
 import { AlertTypeModel } from '../../../../../triggers_actions_ui/public/types';
 import { LOG_DOCUMENT_COUNT_ALERT_TYPE_ID } from '../../../../common/alerting/logs/types';
-import { ExpressionEditor } from './expression_editor';
 import { validateExpression } from './validation';
 
 export function getAlertType(): AlertTypeModel {
@@ -17,7 +17,7 @@ export function getAlertType(): AlertTypeModel {
       defaultMessage: 'Log threshold',
     }),
     iconClass: 'bell',
-    alertParamsExpression: ExpressionEditor,
+    alertParamsExpression: React.lazy(() => import('./expression_editor/editor')),
     validate: validateExpression,
     defaultActionMessage: i18n.translate(
       'xpack.infra.logs.alerting.threshold.defaultActionMessage',
