+++ x-pack/plugins/infra/public/lib/lib.ts
@@ -4,101 +4,17 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { IModule, IScope } from 'angular';
-import { NormalizedCacheObject } from 'apollo-cache-inmemory';
-import ApolloClient from 'apollo-client';
-import { AxiosRequestConfig } from 'axios';
-import React from 'react';
-import { Observable } from 'rxjs';
-import * as rt from 'io-ts';
 import { i18n } from '@kbn/i18n';
-import { SourceQuery } from '../graphql/types';
+import * as rt from 'io-ts';
 import {
-  SnapshotMetricInput,
-  SnapshotGroupBy,
   InfraTimerangeInput,
+  SnapshotGroupBy,
+  SnapshotMetricInput,
   SnapshotNodeMetric,
   SnapshotNodePath,
 } from '../../common/http_api/snapshot_api';
+import { SourceQuery } from '../graphql/types';
 import { WaffleSortOption } from '../pages/metrics/inventory_view/hooks/use_waffle_options';
-
-export interface InfraFrontendLibs {
-  apolloClient: InfraApolloClient;
-  observableApi: InfraObservableApi;
-}
-
-export type InfraTimezoneProvider = () => string;
-
-export type InfraApolloClient = ApolloClient<NormalizedCacheObject>;
-
-export interface InfraFrameworkAdapter {
-  // Insstance vars
-  appState?: object;
-  kbnVersion?: string;
-  timezone?: string;
-
-  // Methods
-  setUISettings(key: string, value: any): void;
-  render(component: React.ReactElement<any>): void;
-  renderBreadcrumbs(component: React.ReactElement<any>): void;
-}
-
-export type InfraFramworkAdapterConstructable = new (
-  uiModule: IModule,
-  timezoneProvider: InfraTimezoneProvider
-) => InfraFrameworkAdapter;
-
-// TODO: replace AxiosRequestConfig with something more defined
-export type InfraRequestConfig = AxiosRequestConfig;
-
-export interface InfraApiAdapter {
-  get<T>(url: string, config?: InfraRequestConfig | undefined): Promise<T>;
-  post(url: string, data?: any, config?: AxiosRequestConfig | undefined): Promise<object>;
-  delete(url: string, config?: InfraRequestConfig | undefined): Promise<object>;
-  put(url: string, data?: any, config?: InfraRequestConfig | undefined): Promise<object>;
-}
-
-export interface InfraObservableApiPostParams<RequestBody extends {} = {}> {
-  url: string;
-  body?: RequestBody;
-}
-
-export type InfraObservableApiResponse<BodyType extends {} = {}> = Observable<{
-  status: number;
-  response: BodyType;
-}>;
-
-export interface InfraObservableApi {
-  post<RequestBody extends {} = {}, ResponseBody extends {} = {}>(
-    params: InfraObservableApiPostParams<RequestBody>
-  ): InfraObservableApiResponse<ResponseBody>;
-}
-
-export interface InfraUiKibanaAdapterScope extends IScope {
-  breadcrumbs: any[];
-  topNavMenu: any[];
-}
-
-export interface InfraKibanaUIConfig {
-  get(key: string): any;
-  set(key: string, value: any): Promise<boolean>;
-}
-
-export interface InfraKibanaAdapterServiceRefs {
-  config: InfraKibanaUIConfig;
-  rootScope: IScope;
-}
-
-export type InfraBufferedKibanaServiceCall<ServiceRefs> = (serviceRefs: ServiceRefs) => void;
-
-export interface InfraField {
-  name: string;
-  type: string;
-  searchable: boolean;
-  aggregatable: boolean;
-}
-
-export type InfraWaffleData = InfraWaffleMapGroup[];
 
 export interface InfraWaffleMapNode {
   pathId: string;
@@ -220,8 +136,6 @@ export interface InfraOptions {
   timerange: InfraTimerangeInput;
   wafflemap: InfraWaffleMapOptions;
 }
-
-export type Omit<T1, T2> = Pick<T1, Exclude<keyof T1, keyof T2>>;
 
 export interface InfraWaffleMapBounds {
   min: number;
